requirements.txt
absl-py==0.13.0
astunparse==1.6.3
autopep8==1.5.7
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
cymem==2.0.5
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.1.0/en_core_web_lg-3.1.0-py3-none-any.whl
filelock==3.0.12
Flask==2.0.2
flatbuffers==1.12
gast==0.4.0
gensim==4.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
huggingface-hub==0.0.16
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
keybert==0.4.0
Markdown==3.3.4
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pathy==0.6.0
Pillow==8.3.2
preshed==3.0.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydantic==1.8.2
pyemd==0.5.1
Pygments==2.10.0
pyparsing==2.4.7
python-dotenv==0.19.0
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.9.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.15.0
smart-open==5.2.1
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
thinc==8.0.9
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
transformers==4.10.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
wrapt==1.12.1
